pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyword-report"
version = "0.1.0"
description = "Keyword opportunity report generator for local service businesses"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "weasyprint>=62.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
keyword-report = "keyword_report.cli:main"
